backend for functionalities

+++ Bookstore/src/main/java/com/example/Bookstore/dto/SaleDTO.java
@@ -15,7 +15,7 @@ import java.util.List;
 @Builder
 @ToString
 public class SaleDTO {
-    private int sum;
+    private double sum;
     private LocalDate date;
     private User user;
     private List<Book> books;
